Fix scientific notation in frequency display and add coding standards - v2.49.1
Fixed:
- EditQSODialog: Frequency field now displays as integer (e.g., "7074000")
  instead of scientific notation (e.g., "7.074e+06")
- Cast to qint64 when converting freq_t to QString to force integer formatting

Added:
- docs/CODING_STANDARDS.md - Project coding standards document
  * Rule: NEVER use scientific notation for user-facing number displays
  * Examples of correct frequency, serial number, and zone formatting
  * Qt-specific guidelines for QString number conversion
  * Amateur radio conventions (callsign handling, RST reports, frequencies)
  * Dialog design patterns and model-view separation guidelines

This ensures ham radio operators see frequencies and other numeric values
in the expected integer format, not confusing scientific notation.

+++ src/ui/dialogs/EditQSODialog.cpp
@@ -195,7 +195,8 @@ void EditQSODialog::loadQSOData() {
     // Basic fields
     m_timestampEdit->setDateTime(m_qso.timestamp);
     m_callsignEdit->setText(m_qso.callsign);
-    m_frequencyEdit->setText(QString::number(m_qso.frequency));
+    // Format frequency as integer (never use scientific notation)
+    m_frequencyEdit->setText(QString::number(static_cast<qint64>(m_qso.frequency)));
 
     // Set mode combo
     int modeIndex = m_modeCombo->findData(static_cast<int>(m_qso.mode));
